Rename rostrum to lineup for clarity
The attendance timeline section was called "rostrum" which was unclear.
Renamed the route from /rostrum to /lineup, moved components from
components/rostrum/ to components/lineup/, and updated all navigation
labels and documentation references.
