+++ components/layout/DesktopHeader.tsx
@@ -7,7 +7,7 @@ import { cn } from "@/lib/utils/cn";
 
 const navLinks = [
   { label: "Hub", href: "/" },
-  { label: "Rostrum", href: "/rostrum" },
+  { label: "Lineup", href: "/lineup" },
   { label: "Feasts", href: "/feasts" },
   { label: "Crew", href: "/crew" },
   { label: "Basecamp", href: "/basecamp" },
