+++ components/layout/MobileTabBar.tsx
@@ -16,8 +16,8 @@ const tabs = [
     ),
   },
   {
-    label: "Rostrum",
-    href: "/rostrum",
+    label: "Lineup",
+    href: "/lineup",
     icon: (
       <svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
         <rect x="3" y="4" width="18" height="18" rx="2" ry="2" />
