+++ components/lineup/EditTripModal.tsx
