+++ components/lineup/TimelineCell.tsx
